Add files via upload
The foundation of this study is a comprehensive experimental dataset meticulously compiled by John-Alan Pascoe as part of his doctoral research (J. A. Pascoe 2016). This dataset (J. A. Pascoe, Alderliesten, and Benedictus 2016) originates from a series of quasi-static and fatigue crack growth experiments conducted on DCB specimens. These specimens are a standard and critical methodology for assessing the crack propagation resistance and long-term durability of adhesive bonds under sustained loads.
For more details check the DOI links below:
https://doi.org/10.4233/uuid:ebbf552a-ce98-4ab6-b9cc-0b939e12ba8b
https://doi.org/10.4121/UUID:AC105275-9DD6-4846-841D-4B0F164E6503